Updating gulp, webpack, and Dockerfile
Unstable commit; committing to track and save progress.

Add Dockerfile to .dockerignore to prevent changes to the Dockfile itself from triggering rebuild of earlier steps of the Dockerfile

Move to Ubuntu 18 for Dockerfile instead of 16. Not yet moving to Ubuntu 20 as Node 8 is not supported on Ubuntu 20.

Rewrote Dockerfile as a multi-stage build to both prevent constant rebuild of everything on small changes and also to only copy needed files into the end runtime.

Moved a number of modules from bower.json to package.json as updated webpack was not able to recognize those specific modules and do dynamic module resolution correctly with them unless the npm version of those modules are used.

Moved to using a different lato font package as the specific module was/is not on npm.

Split webpack config up into two files as it was done in a non-standard looking way to begin with. Now the webpack config file contains only the standard looking webpack config as it should.

Updated gulp to a new version and rewrote the old incompatible bits of the gulpfile to work with the new version.

Temporarily removed deletion of /res/build from clean step of the gulpfile.

Corrected hardcoded module references to modules that are now in node_modules/@devicefarmer

Moved prepublish step of package.json into Dockerfile. This could confuse users not using docker though and may have to be changed back.

Updated to a new version of webpack and altered the webpack file to work with the new way. Angular is still broken and will need to be updated. The old angular method in webpack doesn't work with the new webpack version. TODO

Using versions of less-loader and sass-loader a major version behind as the latest versions require node 10.

Added '-loader' to all the various webpack loaders used not only in webpack config but also in a few module require lines.

Removed the commons.entry.js script lines as the new webpack build doesn't create a commons script.

Attempted to fix the way ngRoute/app is loaded by angular in auth/mock. Still isn't working. Am going to switch to newer angular...

Ran lang/translations json files through a JSON pretty printer as it looked horrible and lack of spaces is not user / curious person friendly.

webpack config file still in progress of changes.