Continuing to get updated build/frontend working
Updated Dockerfile to use even more stages to separate frontend build from backend build.
The 'res' folder has mixed stuff in it for both the frontend and backend so this process is still not complete.

Moved all the angular modules into their own grouping in bower.json for readability. Previously attempted to move all the angular modules into npm package.json but that failed so they have to remain here in bower.json for now.

Updated angular to version 1.8.0. This works "for the most part" with various things broken here and there that am in process of fixing.

Added ng-annotate to webpack process, so as to use ngInject to get some dependencies running in places where old webpack 2.0 process was handling it. Updated a few places that need njInject annotation so far.

Added missing d3 import into app.js

Updated signinController to use `then` instead of deprecated ( and no longer working ) success and error functions.

Continued to update webpack config so that it has what is needed to startup the frontend using webpack 4. Login screen works now but is missing styling.

Main page after login is currently broken on angular-hotkeys and/or mousetrap. That is being worked on.