Alert response functionality
The advanced section has a new button now to show a dialog for responding to an on-screen alert.

Initial mouse wheel support for scrolling screen vertically has been added, but it has issues. If you scroll more than a single "notch" at a time it can freeze the WDA.

For now have added in console logging of messages via the websocket to the backend. This makes events far easier to debug and also generally see what is happening.